Add files statistics_example_wise_common.h and statistics_label_wise_common.h.
